requirements.txt
Django == 2.1.3
django-bootstrap3 == 11.0.0
GitPython == 2.1.11